+++ libretro-common/vfs/vfs_implementation.c
@@ -193,8 +193,8 @@ libretro_vfs_implementation_file *retro_vfs_file_open_impl(const char *path, uns
 
 #ifdef VFS_FRONTEND
    const char                 *dumb_prefix  = "vfsonly://";
-   int                      dumb_prefix_len = strlen(dumb_prefix);
-   int                             path_len = strlen(path);
+   int                      dumb_prefix_len = (int)strlen(dumb_prefix);
+   int                             path_len = (int)strlen(path);
 
    if (path_len >= dumb_prefix_len)
       if (!memcmp(path, dumb_prefix, dumb_prefix_len))
